Fix arg passing bug
Using ...arguments was causing an error where WinstonLoggerAdapter would
not configure since the args were an array.  While this could have been
solved to continue allowing a variable number of arguments, I elected
to take the easy route and only accept a single argument.

To create the logging adapter, one can simply pass a string with the stream
name or they can fully configure by passing in an object.